Moved the original code to a copy of the app file (App (original).jsx). Added a note with all my tasks to a comment in App.jsx
